updated javadoc and validated code changes
