+++ javadoc/constant-values.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_162) on Mon Oct 29 21:43:32 EDT 2018 -->
+<!-- Generated by javadoc (1.8.0_162) on Mon Nov 12 21:53:54 EST 2018 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <title>Constant Field Values</title>
-<meta name="date" content="2018-10-29">
+<meta name="date" content="2018-11-12">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
 </head>
